+++ src/features/tts/commands/mod.rs
@@ -1,0 +1,14 @@
+pub use join::join;
+pub use leave::leave;
+pub use set_nick::set_nick;
+pub use set_model::set_model;
+pub use set_speaker::set_speaker;
+pub use view_models::view_models;
+
+pub mod join;
+pub mod leave;
+pub mod set_nick;
+pub mod set_model;
+pub mod set_speaker;
+pub mod view_models;
+
